requirement.txt
python-dotenv
langextract
"langextract[openai]" 
